requirements.txt
requests==2.26.0
python-dotenv==0.19.1
PyYAML==6.0
cryptography>=41.0.0
websockets>=12.0
pytest>=7.0.0
pytest-mock>=3.10.0